don't temporarily use PAUSED state
This also reverts the previous "fix", for a situation where decoding two files
immediately adjacently would cause a hang. That "fix" did make the problem go
away, but it introduced a new problem in which files would stay open rather than
being properly closed at the end of a decode. (This was probably due to the
removed decodebin no longer being shut down because there wasn't a reference to
it from the pipeline.)

Anyway, a new fix was needed for the hang problem. The hang would always occur
when trying to transition from PAUSED to PLAYING state when the pads were
connected up. I'm not sure why I ever used the PAUSED state; going straight to
PLAYING in the constructor seems to work just fine (and solves the same problem
as the previous "fix").