fix occasional gstreamer hang
The backend would hang unrecoverably in the buffer dequeue; this change seems to
have fixed it. I believe that the problem was caused by the fact that buffers
could be decoded before the "playing" flag was set on the object, leading to
missed buffers when Gstreamer ran fast.
- The main thread started the pipeline and acquired the semaphore.
- The file was decoded very quickly, and _new_buffer was called repeatedly but
  did nothing because self.playing was False.
- Finally, when self.playing became True, the file had already been decoded, so
  the queue was empty. Thus the self.queue.get() blocked forever.
It was a dumb idea to make the modification of the "playing" flag unsynchronized
with respect to its accesses in _pad_added.