Fixes `subprocess` exception on WSL
This fixes issue #51 regarding `self.proc.kill()` when running on Windows Subsystem for Linux.
On WSL, ffmpeg is terminating normally, but `returncode` isn't being updated.  Calling `poll()` on the subprocess returns immediately and forces an update of `returncode`.  This allows the code to skip the `kill()` call if the process is already done.