Separate out backend detection from the audio_open() method.
The backend detection can be inefficient. For example, detecting
CoreAudio may lead to 'ctypes' invoking a C compiler in order to
search for the library.

If a process calls audio_open() many times, it will be more efficient
if backend detection only happens once. This is now possible as there
is a new available_backend() method which you can call once and then
save the result. The list of backends can be passed to audio_open()
in a new 'backends' keyword arg.

Fixes https://github.com/beetbox/audioread/issues/81